保存未提交的更改

+++ routes/web.php
@@ -17,15 +17,20 @@ use Illuminate\Support\Facades\Route;
     return view('populationEducations');
 });*/
 
+
 Route::get('/', [PopulationEducationsController::class,'index']);
 Route::get('/about', [PopulationEducationsController::class,'about'])->name('about');
 
 Route::get('/{id}',[PopulationEducationsController::class,'show'])->where('id','[0-9]+')->name('region.show');
 Route::get('/{id}/edit',[PopulationEducationsController::class,'edit'])->where('id','[0-9]+')->name('region.edit');
 Route::patch('/update/{id}',[PopulationEducationsController::class,'update'])->where('id','[0-9]+')->name('region.update');
-Route::delete('/delete/{id}',[PopulationEducationsController::class,'destroy'])->where('id','[0-9]+')->name('region.destroy');
-Route::get('/region/create',[PopulationEducationsController::class,'create'])->name('region.create');
-Route::post('/region/store',[PopulationEducationsController::class,'store'])->name('region.store');
+Route::delete('/delete/{id}',[PopulationEducationsController::class,'destroy'])->where('id','[0-9]+')->name('region.destroy')->middleware('can:admin');
+Route::get('/region/create',[PopulationEducationsController::class,'create'])->name('region.create')->middleware('can:admin');
+Route::post('/region/store',[PopulationEducationsController::class,'store'])->name('region.store')->middleware('can:admin');
 
 
 
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
